Update README on how to override default values
Currently when trying to override the global ts-jest field with custom values,
the README.md points towards the local configuration that the module uses. This
version of the configuration will not work if the configuration is in any other
location other than in the preset module.

Related: https://github.com/thymikee/jest-preset-angular/issues/286#issuecomment-519050423

For a user to override this part of the object the user must specify the location
of the preset module for it to resolve correctly.

Added a clear example so users will know how to override preset configuration
values and mix them with normal Jest configuration values.