refactor(config): use `readConfiguration` from `@angular/compiler-cli`
Introduce `NgJestConfig` which extends `ts-jest` config class `ConfigSet` and use `readConfiguration` to read tsconfig and get Angular compiler configs.

With `readConfiguration` we are assured that we have similar tsconfig as well as Angular compiler options like Angular CLI does. Later this config will be required to use with Angular compiler.