feat: add experimental precompilation to global setup
Precompilation uses TypeScript API + Angular AST transformers to compile all files which are included in the tsconfig file one is using for Jest.

With precompilation, the precompiled output is written in disk and `ng-jest-transformer` will read disk to get it. After that `ng-jest-transformer` can reuse the precompilation output to reduce the resources to compile codes as well as fixing some issues related to missing source files in compiler.

## Caveats

TypeScript solution style is not supported yet. To workaround it, one needs to specify a wild card syntax in the tsconfig file which is provided to `ngJest` option.